+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "rsschool-react",
   "private": true,
-  "version": "0.0.1",
+  "version": "0.0.2",
   "homepage": "https://K98940.github.io/rsschool-react",
   "type": "module",
   "scripts": {
