+++ src/components/header/header.module.css
@@ -65,6 +65,7 @@
 }
 
 .btnSearch {
+  flex: 0 0 auto;
   padding-inline: 0.5em;
 
   font-size: inherit;
