+++ src/components/index/index.module.css
@@ -1,0 +1,7 @@
+.container {
+  flex: 1 1 60%;
+}
+
+.title {
+  opacity: 0.2;
+}
