+++ src/components/index/index.tsx
@@ -1,9 +1,9 @@
 import classes from './index.module.css';
 
-export default function Index() {
+export const Index = () => {
   return (
     <section className={classes.container}>
-      <h1 className={classes.title}>Select an episode...</h1>
+      <h1 className={classes.title}>Select episode...</h1>
     </section>
   );
-}
+};
