+++ src/components/layout/layout.tsx
@@ -1,0 +1,75 @@
+import api from '@/api/api';
+import { Main } from '../main/main';
+import Header from '../header/header';
+import { Status } from '@/types/types';
+import Spinner from '../spinner/spinner';
+import { ls } from '@/helpers/localStorage';
+import { initialState } from '@/helpers/constants';
+import { useCallback, useEffect, useState } from 'react';
+import { isEpisodeBaseResponse } from '@/helpers/predicates';
+
+export const Layout = () => {
+  const [state, setState] = useState(initialState);
+  const [page, setPage] = useState(0);
+
+  const handleChangePage = (page: number): void => {
+    setPage(page);
+  };
+
+  const updateStatus = (newStatus: Status): void => {
+    setState((prevState) => ({ ...prevState, status: newStatus }));
+  };
+
+  const handleQueryChange = (e: React.ChangeEvent<HTMLInputElement>): void => {
+    setState({ ...state, query: e.target.value });
+  };
+
+  const handleSearch = (): void => {
+    if (state.status === 'submitting') return;
+    getData(state.query);
+  };
+
+  const handleQueryReset = (): void => {
+    setState((prevState) => ({ ...prevState, query: '' }));
+  };
+
+  const getData = useCallback(
+    (term: string) => {
+      updateStatus('submitting');
+      api
+        .searchEpisodes(term.trim(), page)
+        .then((resp) => {
+          if (isEpisodeBaseResponse(resp)) {
+            setState((prevState) => ({ ...prevState, data: resp }));
+          } else {
+            console.log(`this is a bad response (code: ${resp.status}) `, resp.statusText);
+          }
+        })
+        .catch(() => console.log('this is a network error'))
+        .finally(() => {
+          updateStatus('idle');
+        });
+    },
+    [page],
+  );
+
+  useEffect(() => {
+    const term = ls.get();
+    term && setState((prevState) => ({ ...prevState, query: term }));
+    getData(term || '');
+  }, [getData]);
+
+  return (
+    <>
+      {state.status === 'submitting' && <Spinner />}
+      <Header
+        query={state.query}
+        handleSearch={handleSearch}
+        handleQueryChange={handleQueryChange}
+        handleQueryReset={handleQueryReset}
+        status={state.status}
+      />
+      <Main data={state.data} handleChangePage={handleChangePage} />
+    </>
+  );
+};
