+++ src/components/main/main.module.css
@@ -1,7 +1,7 @@
 .main {
   display: flex;
   flex: 1 1 auto;
-  max-height: calc(100% - 5em);
+  max-height: calc(100% - 8em);
   gap: 1em;
   color: black;
   padding: 1em;
@@ -14,6 +14,7 @@
 .episodesNavigation {
   block-size: 100%;
   min-inline-size: 10em;
+  margin-block-end: 1em;
   overflow-y: auto;
   overflow-x: hidden;
