--- src/routes/index/index.module.css
@@ -1,12 +1,0 @@
-.container {
-  display: flex;
-  flex: 1 1 60%;
-  block-size: 100%;
-  align-items: center;
-  justify-content: center;
-}
-
-.title {
-  opacity: 0.2;
-  transform: translateY(-100%);
-}
